# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/order/{id}/orderline": {
            "get": {
                "description": "Get all Orderline by OrderId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get Orderline.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Orderline"
                            }
                        }
                    }
                }
            }
        },
        "/orderline": {
            "get": {
                "description": "Get a list of Orderline.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderline"
                ],
                "summary": "Get all Orderline.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Orderline"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Orderline.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderline"
                ],
                "summary": "Create New Orderline.",
                "parameters": [
                    {
                        "description": "the body to create a new Orderline",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.orderlineInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Orderline"
                        }
                    }
                }
            }
        },
        "/orderline/{id}": {
            "get": {
                "description": "Get an Orderline by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderline"
                ],
                "summary": "Get Orderline.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orderline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Orderline"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Orderline by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderline"
                ],
                "summary": "Delete one Orderline.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orderline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Orderline by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orderline"
                ],
                "summary": "Update Orderline.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orderline id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age rating category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.orderlineInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Orderline"
                        }
                    }
                }
            }
        },
...
        "/product-categories": {
            "get": {
                "description": "Get a list of ProductCategory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Get all ProductCategory.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProductCategory"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new ProductCategory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Create New ProductCategory.",
                "parameters": [
                    {
                        "description": "the body to create a new ProductCategory",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.productCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    }
                }
            }
        },
        "/product-categories/{id}": {
            "get": {
                "description": "Get an ProductCategory by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Get ProductCategory.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductCategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ProductCategory by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Delete one ProductCategory.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductCategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update ProductCategory by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Update ProductCategory.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductCategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update product category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.productCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductCategory"
                        }
                    }
                }
            }
        },
        "/product-categories/{id}/product": {
            "get": {
                "description": "Get all Product by ProductCategoryId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Get Product.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProductCategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Products"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.orderlineInput": {
            "type": "object",
            "properties": {
                "orderID": {
                    "type": "integer"
                },
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "controllers.productCategoryInput": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                }
            }
        },
...
        "models.Orderline": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderID": {
                    "type": "integer"
                },
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ProductCategory": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Products": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "product_category_ID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}